pyproject.toml
[project]
name = "baby-weight-recorder"
version = "0.1.0"
description = "嬰兒體重紀錄系統 - 記錄並追蹤嬰兒體重成長"
readme = "README.md"
requires-python = ">=3.12"
license = { text = "MIT" }
authors = [{ name = "Baby Weight Team" }]
keywords = ["baby", "weight", "health", "tracking"]

dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "pydantic[email]>=2.10.0",
    "pydantic-settings>=2.6.0",
    "google-cloud-firestore>=2.19.0",
    "google-cloud-secret-manager>=2.20.0",
    "python-jose[cryptography]>=3.3.0",
    "passlib[bcrypt]>=1.7.4",
    "httpx>=0.28.0",
    "python-ulid>=2.2.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3.0",
    "pytest-asyncio>=0.24.0",
    "pytest-cov>=6.0.0",
    "ruff>=0.8.0",
    "mypy>=1.13.0",
    "pre-commit>=4.0.0",
    "httpx>=0.28.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["api", "auth"]

[tool.ruff]
target-version = "py312"
line-length = 100
exclude = [
    ".git",
    ".mypy_cache",
    ".ruff_cache",
    ".venv",
    "__pycache__",
    "build",
    "dist",
]

[tool.ruff.lint]
select = [
    "E",      # pycodestyle errors
    "W",      # pycodestyle warnings
    "F",      # Pyflakes
    "I",      # isort
    "B",      # flake8-bugbear
    "C4",     # flake8-comprehensions
    "UP",     # pyupgrade
    "ARG",    # flake8-unused-arguments
    "SIM",    # flake8-simplify
    "TCH",    # flake8-type-checking
    "PTH",    # flake8-use-pathlib
    "RUF",    # Ruff-specific rules
]
ignore = [
    "E501",   # line too long (handled by formatter)
    "B008",   # do not perform function calls in argument defaults (FastAPI Depends)
    "RUF001", # ambiguous unicode characters in strings (中文字符)
    "RUF002", # ambiguous unicode characters in docstrings (中文標點符號)
    "RUF003", # ambiguous unicode characters in comments (中文標點符號)
    "RUF012", # mutable class attributes should use ClassVar
    "RUF022", # __all__ not sorted (prefer logical grouping)
    "ARG001", # unused function argument (often needed for interface compatibility)
    "UP046",  # generic class uses Generic subclass (Python 3.12 style not always desired)
]

[tool.ruff.lint.isort]
known-first-party = ["api", "auth"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
skip-magic-trailing-comma = false
line-ending = "auto"

[tool.mypy]
python_version = "3.12"
# 放寬 strict mode，待之後逐步修復
strict = false
warn_return_any = false
warn_unused_ignores = false
disallow_untyped_defs = false
disallow_incomplete_defs = false
check_untyped_defs = true
disallow_untyped_decorators = false
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_configs = true
show_error_codes = true
show_column_numbers = true
pretty = true
ignore_missing_imports = true
explicit_package_bases = true
namespace_packages = true

[[tool.mypy.overrides]]
module = [
    "google.cloud.*",
    "passlib.*",
    "jose.*",
    "ulid.*",
]
ignore_missing_imports = true

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "function"
addopts = [
    "-v",
    "--tb=short",
    "--strict-markers",
]
markers = [
    "e2e: End-to-end tests that require external services (Docker/Podman)",
    "unit: Unit tests",
    "integration: Integration tests",
    "e2e: End-to-end tests",
]

[tool.coverage.run]
source = ["api", "auth"]
branch = true
omit = ["*/tests/*", "*/__init__.py"]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise NotImplementedError",
    "if TYPE_CHECKING:",
    "if __name__ == .__main__.:",
]
